requirements.txt
discord.py==1.6.0
asyncio==3.4.3